split search component
